+++ app/pipelines/pipelines/requestor.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Pipelines
+  # Requestor
+  module Requestor
+  end
+end
